refactor(parser): Decouple parsing from state
