+++ crates/toml_edit/src/document.rs
@@ -70,7 +70,7 @@ impl FromStr for Document {
 
     /// Parses a document from a &str
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        parser::TomlParser::parse(s.as_bytes())
+        parser::document(s.as_bytes())
     }
 }
