+++ crates/toml_edit/src/parser/mod.rs
@@ -10,147 +10,20 @@ mod errors;
 mod inline_table;
 mod key;
 pub(crate) mod numbers;
+mod state;
 pub(crate) mod strings;
 mod table;
 mod trivia;
 mod value;
 
+pub(crate) use self::document::document;
 pub use self::errors::TomlError;
 pub(crate) use self::key::is_unquoted_char;
 pub(crate) use self::key::key as key_path;
 pub(crate) use self::key::simple_key;
 pub(crate) use self::value::value as value_parser;
 
-use self::table::duplicate_key;
-use crate::key::Key;
-use crate::parser::errors::CustomError;
-use crate::repr::Decor;
-use crate::{ArrayOfTables, Document, Entry, Item, Table};
-
-pub(crate) struct TomlParser {
-    document: Document,
-    trailing: String,
-    current_table_position: usize,
-    current_table: Table,
-    current_is_array: bool,
-    current_table_path: Vec<Key>,
-}
-
-impl TomlParser {
-    pub(crate) fn start_aray_table(
-        &mut self,
-        path: Vec<Key>,
-        decor: Decor,
-    ) -> Result<(), CustomError> {
-        debug_assert!(!path.is_empty());
-        debug_assert!(self.current_table.is_empty());
-        debug_assert!(self.current_table_path.is_empty());
-
-        // Look up the table on start to ensure the duplicate_key error points to the right line
-        let root = self.document.as_table_mut();
-        let parent_table = Self::descend_path(root, &path[..path.len() - 1], false)?;
-        let key = &path[path.len() - 1];
-        let entry = parent_table
-            .entry_format(key)
-            .or_insert(Item::ArrayOfTables(ArrayOfTables::new()));
-        entry
-            .as_array_of_tables()
-            .ok_or_else(|| duplicate_key(&path, path.len() - 1))?;
-
-        self.current_table_position += 1;
-        self.current_table.decor = decor;
-        self.current_table.set_position(self.current_table_position);
-        self.current_is_array = true;
-        self.current_table_path = path;
-
-        Ok(())
-    }
-
-    pub(crate) fn start_table(&mut self, path: Vec<Key>, decor: Decor) -> Result<(), CustomError> {
-        debug_assert!(!path.is_empty());
-        debug_assert!(self.current_table.is_empty());
-        debug_assert!(self.current_table_path.is_empty());
-
-        // 1. Look up the table on start to ensure the duplicate_key error points to the right line
-        // 2. Ensure any child tables from an implicit table are preserved
-        let root = self.document.as_table_mut();
-        let parent_table = Self::descend_path(root, &path[..path.len() - 1], false)?;
-        let key = &path[path.len() - 1];
-        if let Some(entry) = parent_table.remove(key.get()) {
-            match entry {
-                Item::Table(t) if t.implicit => {
-                    self.current_table = t;
-                }
-                _ => return Err(duplicate_key(&path, path.len() - 1)),
-            }
-        }
-
-        self.current_table_position += 1;
-        self.current_table.decor = decor;
-        self.current_table.set_position(self.current_table_position);
-        self.current_is_array = false;
-        self.current_table_path = path;
-
-        Ok(())
-    }
-
-    pub(crate) fn finalize_table(&mut self) -> Result<(), CustomError> {
-        let mut table = std::mem::take(&mut self.current_table);
-        let path = std::mem::take(&mut self.current_table_path);
-
-        let root = self.document.as_table_mut();
-        if path.is_empty() {
-            assert!(root.is_empty());
-            std::mem::swap(&mut table, root);
-        } else if self.current_is_array {
-            let parent_table = Self::descend_path(root, &path[..path.len() - 1], false)?;
-            let key = &path[path.len() - 1];
-
-            let entry = parent_table
-                .entry_format(key)
-                .or_insert(Item::ArrayOfTables(ArrayOfTables::new()));
-            let array = entry
-                .as_array_of_tables_mut()
-                .ok_or_else(|| duplicate_key(&path, path.len() - 1))?;
-            array.push(table);
-        } else {
-            let parent_table = Self::descend_path(root, &path[..path.len() - 1], false)?;
-            let key = &path[path.len() - 1];
-
-            let entry = parent_table.entry_format(key);
-            match entry {
-                Entry::Occupied(entry) => {
-                    match entry.into_mut() {
-                        // if [a.b.c] header preceded [a.b]
-                        Item::Table(ref mut t) if t.implicit => {
-                            std::mem::swap(t, &mut table);
-                        }
-                        _ => return Err(duplicate_key(&path, path.len() - 1)),
-                    }
-                }
-                Entry::Vacant(entry) => {
-                    let item = Item::Table(table);
-                    entry.insert(item);
-                }
-            }
-        }
-
-        Ok(())
-    }
-}
-
-impl Default for TomlParser {
-    fn default() -> Self {
-        Self {
-            document: Document::new(),
-            trailing: String::new(),
-            current_table_position: 0,
-            current_table: Table::new(),
-            current_is_array: false,
-            current_table_path: Vec::new(),
-        }
-    }
-}
+use self::state::ParseState;
 
 #[cfg(test)]
 mod tests {
@@ -624,21 +497,21 @@ key = "value"
             r#"foo = 1979-05-27 # Comment
 "#,
         ];
-        for document in documents {
-            let doc = TomlParser::parse(document.as_bytes());
+        for input in documents {
+            let doc = document(input.as_bytes());
             let doc = match doc {
                 Ok(doc) => doc,
                 Err(err) => {
                     panic!(
                         "Parse error: {}\nFailed to parse:\n```\n{}\n```",
-                        err, document
+                        err, input
                     )
                 }
             };
 
             dbg!(doc.to_string());
-            dbg!(document);
-            assert_eq(document, doc.to_string());
+            dbg!(input);
+            assert_eq(input, doc.to_string());
         }
 
         let parse_only = ["\u{FEFF}
@@ -647,14 +520,14 @@ name = \"foo\"
 version = \"0.0.1\"
 authors = []
 "];
-        for document in parse_only {
-            let doc = TomlParser::parse(document.as_bytes());
+        for input in parse_only {
+            let doc = document(input.as_bytes());
             match doc {
                 Ok(_) => (),
                 Err(err) => {
                     panic!(
                         "Parse error: {}\nFailed to parse:\n```\n{}\n```",
-                        err, document
+                        err, input
                     )
                 }
             }
@@ -662,8 +535,8 @@ authors = []
 
         let invalid_inputs = [r#" hello = 'darkness' # my old friend
 $"#];
-        for document in invalid_inputs {
-            let doc = TomlParser::parse(document.as_bytes());
+        for input in invalid_inputs {
+            let doc = document(input.as_bytes());
 
             assert!(doc.is_err());
         }
